# Implement geo-traits for writing to WKT & perf improvement