--- a/src/geo_traits_to_wkt.rs
+++ b/src/geo_traits_to_wkt.rs
@@ -1,0 +1,3 @@
+use geo_traits::GeometryTrait;
+
+// pub fn to_wkt(g: &impl GeometryTrait) ->