--- a/src/geo_traits_to_wkt.rs
+++ b/src/geo_traits_to_wkt.rs
@@ -1,3 +1,0 @@
-use geo_traits::GeometryTrait;
-
-// pub fn to_wkt(g: &impl GeometryTrait) ->