--- a/src/to_wkt/geo_trait_impl.rs
+++ b/src/to_wkt/geo_trait_impl.rs
@@ -6,6 +6,7 @@ use geo_traits::{
     MultiLineStringTrait, MultiPointTrait, MultiPolygonTrait, PointTrait, PolygonTrait, RectTrait,
     TriangleTrait,
 };
+#[cfg(feature = "geo-types")]
 use geo_types::CoordNum;
 
 use crate::types::{Coord, LineString, Polygon};