--- a/src/to_wkt/geo_trait_impl.rs
+++ b/src/to_wkt/geo_trait_impl.rs
@@ -1,4 +1,4 @@
-use core::fmt;
+use std::fmt;
 use std::fmt::Write;
 
 use geo_traits::{