… src/types/coord.rs
… src/types/linestring.rs
… src/types/multilinestring.rs
… src/types/multipoint.rs
… src/types/multipolygon.rs
… src/types/point.rs
… src/types/polygon.rs