@kylebarron 2024-10-26
All existing tests pass; just a question of whether `geo-traits` should be excluded from default features because it depends on `geo-types`, which is excluded.