@kylebarron 2024-10-29
Well I consider the `todo` (at least for this pr) "I don't know the right way to handle this, I'm open to advice"

Is there a better way to signify this?