@urschrei 2024-10-29
Another alternative / complementary approach is to deny `todo!` in Clippy so a merge becomes gated on no `todo`s