@kylebarron 2024-10-29
Other questions:

- Should we change the existing `geo_types_to_wkt` behavior? That _clones_ geo-types objects into `wkt` objects, and then writes those. Could we deprecate that entire module?
- This is a PR into the other branch. Any opinions on whether to review and merge that PR first?