@kylebarron 2024-11-25
Let's get #123 merged and then we can come back to this to decide how we want to write `Rect`s