@kylebarron 2024-11-26
I updated the Rect writing, changed the `todo!()` to `Error` and this should be good for review now.